+++ src/main/java/com/dnd/backend/user/controller/UserController.java
@@ -2,7 +2,6 @@ package com.dnd.backend.user.controller;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,14 +13,16 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.dnd.backend.user.dto.AddressDTO;
 import com.dnd.backend.user.entity.User;
-import com.dnd.backend.user.service.UserService;
+import com.dnd.backend.user.service.UserServiceImpl;
+
+import lombok.RequiredArgsConstructor;
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/api/user")
 public class UserController {
 
-	@Autowired
-	private UserService userService;
+	private UserServiceImpl userService;
 
 	// 내 정보 조회
 	@GetMapping("/me")
